+++ uu/id.go
@@ -121,18 +121,18 @@ func IDv7WithTime(t time.Time) ID {
 	return id
 }
 
-// V7Time returns the timestamp of a version 7 UUID.
-// If the UUID is not a version 7 UUID, a zero time is returned.
+// V7Time returns the timestamp of a version 7 UUID and true.
+// If the UUID is not a version 7 UUID, a zero time and false are returned.
 //
 // Note that the precision is only milliseconds.
 //
 // See [IDv7WithTime] for the reverse operation.
-func (id ID) V7Time() time.Time {
+func (id ID) V7Time() (t time.Time, ok bool) {
 	if id.Version() != 7 {
-		return time.Time{}
+		return time.Time{}, false
 	}
-	u := *(*uint64)(unsafe.Pointer(&id[0]))              //#nosec G103 -- unsafe OK
-	return time.UnixMilli(int64(u & 0x0000ffffffffffff)) //#nosec G115 -- integer conversion OK
+	u := *(*uint64)(unsafe.Pointer(&id[0]))                    //#nosec G103 -- unsafe OK
+	return time.UnixMilli(int64(u & 0x0000ffffffffffff)), true //#nosec G115 -- integer conversion OK
 }
 
 // IDv7Deterministic returns a version 7 ID with